+++ src/main/java/com/sej/escape/service/theme/ThemeMapper.java
@@ -85,7 +85,7 @@ public class ThemeMapper {
         List<String> areaHierarchy = areaSectionUtil.getTitleFromAreaCode(storeAreaDto.getAreaCode());
         themeDto.setArea(areaHierarchy);
 
-        double starAvg = row[1] != null ? ((BigInteger) row[1]).doubleValue() : 0.0;
+        double starAvg = row[1] != null ? ((BigDecimal) row[1]).doubleValue() : 0.0;
         themeDto.setStar(starAvg);
 
         FileUrlDto file = (FileUrlDto) row[5];
